Logged-in check for voting improvements
Uses UserProfileData to check if the user is currently logged in. If not, the vote button gives an error message when clicked instead of trying to vote.
Makes the querySelector used to find the program vote button more specific, to fix a bug where it occasinally wouldn't be found.
Replaces two vote class constants with an enum.
Adds a space to the updateVoteDisplay to comply with the style guide.