Project nickname fixes (#118)
* Split project button modificaitons into buttons.ts

There was a lot of code surrounding modifiying buttons (replacing the vote button and adding the copy link button) in the project.ts file. Moving this to another file makes it easier to add more code both around buttons and around the project page in the future.

* Update classname and logic for new UI

A KA update removed the "profile-programs" class. Instead, we directly grab the span of the last updated time of the program.