Add check to 404 page for completly hidden programs
If the url of the page looks like a program (we can find a program id), we check if we can find the `#four-oh-four` element, which shows up on KA's 404 page. If we can, we make a network request to the API data for the program, if that doesn't 404, then we know the program actually exsists, and KA is just hiding it from us. We add an element to the page letting the user know whether program is actually deleted or just appears as deleted.