Info for completely hidden programs (#119)
* Add check to 404 page for completely hidden programs

If the URL of the page looks like a program (we can find a program id), we check if we can find the `#four-oh-four` element, which shows up on KA's 404 page. If we can, we make a network request to the API data for the program, if that doesn't 404, then we know the program actually exists, and KA is just hiding it from us. We add an element to the page letting the user know whether program is actually deleted or just appears as deleted.

* Program info table supports completely hidden programs

Changes two table rows (for guardian approved and hidden from hotlist) into one row. Adds a check (to the same row) for if the program is completely hidden (appearsAsDeleted, and most users see a 404 page).
Also adds color coding; red for completely hidden, orange for hidden from hotlist, black for no, and green for guardian approved.