Fixes for KA's new toolbar
The toolbar under projects on KA updated. This introduces fixes for issues with the toggle editor settings button that arose as a result of that.
It is now appened to the left half of the area under the editor, so the pop up has room to expand to the right.
The toggle editor settings button now uses the new styling, matching the Spin Off button.
Add CSS for handling hovering (which KA uses jQuery for) and text styling (KA uses another wrap element).
Removes a spacer that was around this element.
Removes the CSS for the toggle dark theme button that never got removed.
Removes CSS that messed with the .pull-right class as part of respacing elements in this toolbar.