Change how buttons are re-spaced (Fix share button)
The method that was being used to respace buttons was causing issues with the Share button, for reasons I don't fully understand. This fixes those issues by changing button respacing to be done primarily in CSS. We add text nodes to buttons that didn't have one instead of removing these text nodes.
This also changes how the CSS for the report buttons is handled. And fixes a bug where buttons had a dark green outline when hovered over.