Remove profile reply linking & fix flags
This feature added links to a user's replies page, allowing you to jump to a comment. With the discussion update, KA has added a link to comments, so the extension feature is no longer needed (although jumping to replies doesn't work). All of comment-data.ts got removed as part of this.
This includes removing lodash.flatten and zipobject, since they were only used there (this causes the package-lock.json changes).
Creates a discussion.ts file.
This defines functions that get run once the discussion package is loaded. We switch to the Tips and Thanks as the default view instead of questions. Also, for each comment poll the API to get flag data and add them to each comment.
Removes the onReplies function, it's not being used.
Remove functions for the user replies page from the api-utils, and add a function to get data for a comment.