Bug fixes (#168)
* Move profile author hovercards to fire onProgramAboutPage

The hovercard was "lost" if you switched away from and back to the program about tab.

* Remove onNewProgramPage

The onProgramPage no longer requires discussion to be loaded, so it fires on the new program page and the new program page doesn't need a separate check.

* Move editor toggle under program

Moves the toggle editor button under the program, instead of floating above it. (Aside from this looking better IMO, it also) fixes a bug where the CSS for the classes this button had weren't loaded on child account programs or new program pages.

Removes BUTTON_CLASSES from project.ts, which used to hold the button classes used for everything (before the KA UI update). It's no longer used for the hide-editor button, and is unneeded in this file.

Adds a couple more comments explaining CSS rules.