Switch from KAdefine
Get KAID data from window.KA.kaid instead of from loading the module with KAdefine.require.
Fix the check for if you're on your own profile, which was always returning true. Fix the check in UsernameOrKaid.asUsername which was checking if it was a KAID.
Update KAdefine.asyncRequire to throw a depreciation message.
Add getScratchpadUI function to get `window.ScratchpadUI` when/if it loads