Update programdata
Tie to the user link instead of the updated class. Confirmed to work even with child account programs. Unconfirmed with anonymous programs.
Also adds the table a sibling to the program buttons and user info, instead of as a child of the updated element or one higher than whereever it was.
Fixes a bug with the toggle editor settings button where it thought it was an anchor but was previously updated to a button.