Fix class hashes (#179)
* Start fixing class hashes

-Removes BUTTON_CLASSES, used for storing the old class names. We're trying to move away from these because they are unstable.
-Add a "kae-program-button" class to the Copy Link and vote buttons, which is currently unused.
-Get program buttons by `a[role="button"]`
-Move styles for all extension custom buttons into general.css (instead of trying to set a class that KA will match)
-Update the class used to find and place the Editor Settings Toggle button,
-Refactor the Editor Settings Toggle button so that if that class changes, stored settings will still be applied.

* Grab buttons in TS and add a class

Instead of trying to do trash CSS selectors that didn't even work, I grabbed the program buttons in TypeScript and added the ".kae-program-button" class to them. Simplifies the CSS accordingly.

* Refactor buttons

Refactor buttons.ts to export one function, which is responsible for finding the buttons wrap and calling other modifications.

* Update programdata

Tie to the user link instead of the updated class. Confirmed to work even with child account programs. Unconfirmed with anonymous programs.
Also adds the table a sibling to the program buttons and user info, instead of as a child of the updated element or one higher than whereever it was.
Fixes a bug with the toggle editor settings button where it thought it was an anchor but was previously updated to a button.

* Move flag button code into buttons.ts

Move flag button code into buttons.ts, fixing it at the same time to run off the same querySelectorPromise as the other button events.
Removes the flag file!

* Remove old button css

* Move report button from report.ts to buttons.ts

Moved it onto the same querySelectorPromise as everything else!

* Update discussion page check and program button spacing

* Fix lint errors

* Add some discussion buttons listeners

Find all discussion tab buttons (Questions, T&T, Help Requests), and add event listeners to them to trigger looking for new comments.
Find the Load More Comments button and attach a listener to that.

* Add check for replies button & refactor

-Add an event listener to the show replies button
-Move code that handles top-level comments into a specific check
-Add the number of flags to a comment even if there are no flags (makes it clear when the Exension's checking flags)
-Add a bunch of comments

* Refactor load more comments button listener

-We assume that there is at least 1 unalteredComments, because the querySelector fired, and then assume it has a 4 times removed parent (reasonable, if even if we find the wrong element, given the depth of KA's DOM tree). We then grab the last of it's children. If that's a button, we add an event listener, otherwise we exit silently.
-Increase time we search for new comments from 1sec to 2sec, I had it fail loading questions once, and I've cut down on the number times where it shouldn't find comments.

* Listener for answers

-Add an event listener for the button to show answers to questions
-Update error handling to ignore the promise rejection error and still console.error other, legitimate, errors

* Fix lint errors

* 4.6.3

-Remove debug code
-Increase comment searching timer (2s -> 4s)
-Changelog and version bumps to 4.6.3