Move hide editor to Editor Settings
Hide editor is a checkbox inside the Editor Settings popup. Removes code to add it as a toolbar button. Hopefully this more reliable in the future. (Intentionally changes the localStorage key, to avoid the possiblity someone had the editor toggled off and can't figure out how to turn it back on.)

-Removes addProgramAuthorHoverCard. I don't think it's feasible since the Webpack update. We could rewrite it completely or figure out how to get functions working out of webpackJSON, but for now it's removed.

-Moves Toggle Editor Settings button before Oh Noes, to avoid Oh Noes clipping as he changes width. As part of that, relies less on hashed class names